[DOP-19907] - change migrations & move models, mixins
