+++ syncmaster/db/mixins/__init__.py
@@ -1,0 +1,5 @@
+# SPDX-FileCopyrightText: 2023-2024 MTS PJSC
+# SPDX-License-Identifier: Apache-2.0
+from syncmaster.db.mixins.deletable import DeletableMixin
+from syncmaster.db.mixins.resource import ResourceMixin
+from syncmaster.db.mixins.timestamp import TimestampMixin
