+++ syncmaster/db/mixins/resource.py
@@ -1,24 +1,8 @@
 # SPDX-FileCopyrightText: 2023-2024 MTS PJSC
 # SPDX-License-Identifier: Apache-2.0
-from datetime import datetime
 
-from sqlalchemy import BigInteger, Boolean, DateTime, ForeignKey, String, func
+from sqlalchemy import BigInteger, ForeignKey, String
 from sqlalchemy.orm import Mapped, mapped_column
-
-
-class TimestampMixin:
-    created_at: Mapped[datetime] = mapped_column(DateTime, server_default=func.now(), nullable=False)
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime,
-        server_default=func.now(),
-        server_onupdate=func.now(),
-        onupdate=datetime.now,
-        nullable=False,
-    )
-
-
-class DeletableMixin:
-    is_deleted: Mapped[bool] = mapped_column(Boolean, default=False, nullable=False)
 
 
 class ResourceMixin:
