+++ syncmaster/db/models/__init__.py
@@ -1,0 +1,10 @@
+# SPDX-FileCopyrightText: 2023-2024 MTS PJSC
+# SPDX-License-Identifier: Apache-2.0
+from syncmaster.db.models.auth_data import AuthData
+from syncmaster.db.models.base import Base
+from syncmaster.db.models.connection import Connection
+from syncmaster.db.models.group import Group, GroupMemberRole, UserGroup
+from syncmaster.db.models.queue import Queue
+from syncmaster.db.models.run import Run, Status
+from syncmaster.db.models.transfer import Transfer
+from syncmaster.db.models.user import User
